+++ pwnlib/tubes/tube.py
@@ -4,6 +4,7 @@ from __future__ import division
 
 import abc
 import logging
+import os
 import re
 import six
 import string
@@ -867,7 +868,7 @@ class tube(Timeout, Logger):
         is much more usable, since we are using :mod:`pwnlib.term` to print a
         floating prompt.
 
-        Thus it only works in while in :data:`pwnlib.term.term_mode`.
+        Thus it only works while in :data:`pwnlib.term.term_mode`.
         """
 
         self.info('Switching to interactive mode')
@@ -892,13 +893,44 @@ class tube(Timeout, Logger):
         t.daemon = True
         t.start()
 
+        from pwnlib.args import term_mode
         try:
+            os_linesep = os.linesep.encode()
+            to_skip = b''
             while not go.isSet():
                 if term.term_mode:
                     data = term.readline.readline(prompt = prompt, float = True)
+                    if data:
+                        data += self.newline
                 else:
                     stdin = getattr(sys.stdin, 'buffer', sys.stdin)
                     data = stdin.read(1)
+                    # Keep OS's line separator if NOTERM is set and
+                    # the user did not specify a custom newline
+                    # even if stdin is a tty.
+                    if sys.stdin.isatty() and (
+                        term_mode
+                        or context.newline != b"\n"
+                        or self._newline is not None
+                    ):
+                        if to_skip:
+                            if to_skip[:1] != data:
+                                data = os_linesep[: -len(to_skip)] + data
+                            else:
+                                to_skip = to_skip[1:]
+                                if to_skip:
+                                    continue
+                                data = self.newline
+                        # If we observe a prefix of the line separator in a tty,
+                        # assume we'll see the rest of it immediately after.
+                        # This could stall until the next character is seen if
+                        # the line separator is started but never finished, but
+                        # that is unlikely to happen in a dynamic tty.
+                        elif data and os_linesep.startswith(data):
+                            if len(os_linesep) > 1:
+                                to_skip = os_linesep[1:]
+                                continue
+                            data = self.newline
 
                 if data:
                     try: