+++ CHANGELOG.md
@@ -76,6 +76,7 @@ The table below shows which release corresponds to each branch, and what date th
 - [#2144][2144] Removes `p2align 2` `asm()` headers from `x86-32`, `x86-64` and `mips` architectures to avoid inconsistent instruction length when patching binaries
 - [#2177][2177] Support for RISC-V 64-bit architecture
 - [#2186][2186] Enhance `ELF.nx` and `ELF.execstack`
+- [#2129][2129] Handle `context.newline` correctly when typing in `tube.interactive()`
 
 [2185]: https://github.com/Gallopsled/pwntools/pull/2185
 [2062]: https://github.com/Gallopsled/pwntools/pull/2062
@@ -86,6 +87,7 @@ The table below shows which release corresponds to each branch, and what date th
 [2144]: https://github.com/Gallopsled/pwntools/pull/2144
 [2177]: https://github.com/Gallopsled/pwntools/pull/2177
 [2186]: https://github.com/Gallopsled/pwntools/pull/2186
+[2129]: https://github.com/Gallopsled/pwntools/pull/2129
 
 ## 4.10.0 (`beta`)